Move handlers into 'handler' folder
