+++ src/handlers/station-manager-click-handler.js
