+++ src/index.js
@@ -1,5 +1,5 @@
-import menuButtonHandler from "./menu-button-handler.js";
-import stationManagerClickHandler from "./station-manager-click-handler.js";
+import menuButtonHandler from "./handlers/menu-button-handler.js";
+import stationManagerClickHandler from "./handlers/station-manager-click-handler.js";
 
 document.addEventListener("click", menuButtonHandler);
 document.addEventListener("click", stationManagerClickHandler);
